Fix main sitemap population and extract static route helpers
