+++ frontend/server/plugins/sitemap-main-pages.ts
@@ -1,5 +1,4 @@
 import { getRequestURL } from 'h3'
-import type { SitemapUrlInput } from '@nuxtjs/sitemap'
 
 import { APP_ROUTES_SITEMAP_KEY } from '~~/shared/utils/sitemap-config'
 import { getDomainLanguageFromHostname } from '~~/shared/utils/domain-language'
@@ -7,30 +6,8 @@ import { getMainPagePathsForDomainLanguage } from '~~/shared/utils/sitemap-main-
 
 const MAIN_PAGES_SITEMAP_FILENAME = `${APP_ROUTES_SITEMAP_KEY}.xml`
 
-const extractUrlKey = (entry: SitemapUrlInput): string | null => {
-  if (typeof entry === 'string') {
-    return entry
-  }
-
-  if (!entry || typeof entry !== 'object') {
-    return null
-  }
-
-  if ('loc' in entry && typeof entry.loc === 'string' && entry.loc) {
-    return entry.loc
-  }
-
-  if ('url' in entry && typeof (entry as { url?: unknown }).url === 'string') {
-    const url = (entry as { url?: string }).url
-
-    return url && url.length > 0 ? url : null
-  }
-
-  return null
-}
-
 export default defineNitroPlugin((nitroApp) => {
-  nitroApp.hooks.hook('sitemap:input', (ctx) => {
+  nitroApp.hooks.hook('sitemap:sources', (ctx) => {
     if (ctx.sitemapName !== MAIN_PAGES_SITEMAP_FILENAME) {
       return
     }
@@ -38,31 +15,19 @@ export default defineNitroPlugin((nitroApp) => {
     const requestURL = ctx.event ? getRequestURL(ctx.event) : new URL('https://nudger.com')
     const { domainLanguage } = getDomainLanguageFromHostname(requestURL.hostname)
 
-    const staticPaths = getMainPagePathsForDomainLanguage(domainLanguage)
+    const staticPaths = Array.from(new Set(getMainPagePathsForDomainLanguage(domainLanguage)))
 
     if (!staticPaths.length) {
       return
     }
 
-    const seen = new Set<string>()
-
-    ctx.urls.forEach((entry) => {
-      const key = extractUrlKey(entry)
-
-      if (!key) {
-        return
-      }
-
-      seen.add(key)
-    })
-
-    staticPaths.forEach((path) => {
-      if (seen.has(path)) {
-        return
-      }
-
-      ctx.urls.push({ loc: path })
-      seen.add(path)
+    ctx.sources.push({
+      context: {
+        name: 'static-main-pages',
+        description: 'Autodiscovered localized marketing routes',
+      },
+      sourceType: 'user',
+      urls: staticPaths.map((path) => ({ loc: path })),
     })
   })
 })
